requirements.txt
PyQt6==6.6.0
opencv-python==4.8.1.78
sounddevice==0.4.6
pynput==1.7.6
pyautogui==0.9.54
numpy==1.24.3
firebase-admin==6.2.0
python-dotenv==1.0.0
requests>=2.28